+++ webapp/src/components/backstage/playbook_edit/styles.tsx
@@ -19,25 +19,6 @@ export const SidebarBlock = styled.div`
     margin: 0 0 40px;
 `;
 
-export const Setting = styled.div`
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-
-    margin-bottom: 24px;
-`;
-
-export const BackstageGroupToggleHeader = styled.div`
-    font-weight: 600;
-    font-size: 16px;
-    line-height: 24px;
-    color: var(--center-channel-color);
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    column-gap: 12px;
-`;
-
 export const ClockOutline = ({sizePx, color}: {sizePx: number, color?: string}) => (
     <Icon
         path={mdiClockOutline}
