+++ webapp/src/components/backstage/playbook_editor/controls.tsx
@@ -51,7 +51,6 @@ import {
 import {OVERLAY_DELAY} from 'src/constants';
 import {ButtonIcon, PrimaryButton, SecondaryButton, TertiaryButton} from 'src/components/assets/buttons';
 import CheckboxInput from '../runs_list/checkbox_input';
-import StatusBadge, {BadgeType} from 'src/components/backstage/status_badge';
 
 import {displayEditPlaybookAccessModal, openPlaybookRunModal} from 'src/actions';
 import {PlaybookPermissionGeneral} from 'src/types/permissions';
@@ -136,16 +135,6 @@ export const Members = (props: {playbookId: string, numMembers: number}) => {
     );
 };
 
-export const Share = ({playbook: {id}}: ControlProps) => {
-    const dispatch = useDispatch();
-    return (
-        <TertiaryButtonLarger onClick={() => dispatch(displayEditPlaybookAccessModal(id))}>
-            <i className={'icon icon-lock-outline'}/>
-            <FormattedMessage defaultMessage='Share'/>
-        </TertiaryButtonLarger>
-    );
-};
-
 export const CopyPlaybook = ({playbook: {title, id}}: ControlProps) => {
     return (
         <CopyLink
@@ -153,19 +142,6 @@ export const CopyPlaybook = ({playbook: {title, id}}: ControlProps) => {
             to={getSiteUrl() + '/playbooks/playbooks/' + id}
             name={title}
             area-hidden={true}
-        />
-    );
-};
-
-export const ArchivedLabel = ({playbook: {delete_at}}: ControlProps) => {
-    const archived = delete_at !== 0;
-    if (!archived) {
-        return null;
-    }
-    return (
-        <StatusBadge
-            data-testid={'archived-badge'}
-            status={BadgeType.Archived}
         />
     );
 };
